fix(sql): refactor and reuse number type precedence in type system (expr.c)
Previously, there were two conflicting number type orders:

1. sql_highest_type

```
if (a == FIELD_TYPE_NUMBER || b == FIELD_TYPE_NUMBER)
        return FIELD_TYPE_NUMBER;
if (a == FIELD_TYPE_DECIMAL || b == FIELD_TYPE_DECIMAL)
        return FIELD_TYPE_DECIMAL;
if (a == FIELD_TYPE_DOUBLE || b == FIELD_TYPE_DOUBLE)
        return FIELD_TYPE_DOUBLE;
```

2. sql_type_result

```
if (lhs == FIELD_TYPE_NUMBER || rhs == FIELD_TYPE_NUMBER)
        return FIELD_TYPE_NUMBER;
if (lhs == FIELD_TYPE_DOUBLE || rhs == FIELD_TYPE_DOUBLE)
        return FIELD_TYPE_DOUBLE;
if (lhs == FIELD_TYPE_DECIMAL || rhs == FIELD_TYPE_DECIMAL)
        return FIELD_TYPE_DECIMAL;
```

NO_DOC=internal
NO_CHANGELOG=internal

+++ test/sql-luatest/gh_6990_case_operation_type_test.lua
@@ -51,7 +51,7 @@ g.test_case_operation_type = function()
         t.assert_equals(box.execute(sql).metadata[1].type, res)
 
         sql = [[SELECT CASE 1 WHEN 1 THEN -1 WHEN 2 THEN 1.5 ELSE 2e0 END;]]
-        res = "decimal"
+        res = "double"
         t.assert_equals(box.execute(sql).metadata[1].type, res)
 
         sql = [[SELECT typeof(CASE 1 WHEN 1 THEN 1 ELSE {1 : 1} END);]]
